requirements.txt
anthropic>=0.52.0
httpx>=0.28.0
pydantic>=2.11.0
google-api-python-client>=2.170.0
google-auth-oauthlib>=1.2.0
python-dotenv>=1.1.0
pytest>=8.4.0
pytest-asyncio>=1.0.0
